+++ src/bar-graph.js
@@ -4,6 +4,8 @@ var _ = require('underscore')
   , base = require('d3by5-base-chart')
   , barOptions = require('./barOptions')
   , labelOptions = require('./labelOptions')
+  , barPositions = require('./barPositions')
+  , barDimensions = require('./barDimensions')
 ;
 
 module.exports = BarGraph;
@@ -21,6 +23,9 @@ function BarGraph () {
     ANCHOR_BOTTOM: 'bottom',
     ANCHOR_TOP: 'top',
 
+    BARLAYOUT_STACKED: 'stacked',
+    BARLAYOUT_GROUPED: 'grouped',
+
     LABEL_INSIDE: 'inside',
     LABEL_OUTSIDE: 'outside',
     LABEL_FIT: 'fit',
@@ -31,7 +36,8 @@ function BarGraph () {
         anchor: 'right',
         chartClass: 'chart-bar',
         labelPosition: 'none',
-        idPrefix: 'bar-'
+        idPrefix: 'bar-',
+        barLayout: 'grouped'
     },
 
     init: function (selection) {
@@ -44,88 +50,41 @@ function BarGraph () {
     },
 
     draw: function () {
-      var that = this;
-      // global accessible max
-      this.maxValue = d3.max(this.options.data, function (d) {
-                        if (that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL) {
-                          return d3.max(d.values, function (d2) {
-                            return d2.values;
-                          });
-                        }
-                        return d.values;
-      });
-
-      var lengthRange = this.getLengthRange()
-        , breadthRange = this.getBreadthRange()
-        , length
-        , breadth
-        , multiBreadth
-        , _labelOptions
+      var that = this
         , positions = {}
         , dimensions = {}
       ;
+      // force a value for the dataType if there is a multi dimensional dataset
+      this.options.barLayout = (that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL) ? this.options.barLayout || this.BARLAYOUT_GROUPED : null;
 
-      length = d3.scale.linear()
-                        .domain([0, this.maxValue])
-                        .range(lengthRange);
-
-      breadth = d3.scale.ordinal()
-                        .domain(_.map(that.options.data, function (d) {
-                          return d.label;
-                        }))
-                        .rangeRoundBands(breadthRange, 0.1);
-
+      this.maxValue = this.getMaxValue();
+      this.breadthScale = this.getBreadthScale();
+      this.lengthScale = this.getLengthScale();
 
       //
-      // Add a sub category for multi-dimensional data
+      // Create a grouped breadth scale and update the dataset 
       //
       if (that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL) {
-        multiBreadth = d3.scale.ordinal()
-                              .domain(_.map(_.first(that.options.data).values, function (d) {
-                                    return d.label;
-                                  }))
-                              .rangeRoundBands([breadth.rangeBand(), 0]);
+        this.groupedBreadthScale = this.getGroupedBreadthScale();
+
+        if (that.options.barLayout === this.BARLAYOUT_STACKED) {
+          var lpos;
+          _.each(that.options.data, function (data) {
+            lpos = 0;
+            data.values = _.map(data.values, function (d) {
+                            lpos +=  d.values;
+                            d.lpos = lpos;
+                            return d;
+                          });
+          });
+        }
       }
 
 
-
-      //
-      // Set the bar positions to x and y depending on their anchor and the datatype
-      //
-      if (this.isVertical()) {
-        positions.x = function(d) {
-                        return that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL ? multiBreadth(d.label) : 0;
-                      };
-        positions.y = function(d) {
-                        return that.options.anchor === that.ANCHOR_TOP ? that.options.margin.top : that.getCalculatedHeight() - length(d.values);
-                      };
-      } else {
-        positions.x = function(d) {
-                        return that.options.anchor === that.ANCHOR_LEFT ? 0 : that.getCalculatedWidth() - length(d.values);
-                      };
-        positions.y = function(d) {
-                        return that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL ? multiBreadth(d.label) : 0;
-                      };
-      }
-
-      //
-      // Set dimensions
-      //
-      if (this.isVertical()) {
-        dimensions.width = function (d) {
-                              return that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL ? multiBreadth.rangeBand() : breadth.rangeBand();
-                            };
-        dimensions.height = function(d) {
-                              return length(d.values);
-                            };
-      } else {
-        dimensions.width = function (d) {
-                              return length(d.values);
-                            };
-        dimensions.height = function(d) {
-                              return that.options.dataType === that.DATATYPE_MULTIDIMENSIONAL ? multiBreadth.rangeBand() : breadth.rangeBand();
-                            };
-      }
+      positions.x = this.getBarXPos();
+      positions.y = this.getBarYPos();
+      dimensions.width = this.getBarWidth();
+      dimensions.height = this.getBarHeight();
 
       this.selection.each(function() {
 
@@ -150,8 +109,8 @@ function BarGraph () {
             .enter()
             .append('g')
             .attr('transform', function (d) {
-              var _x = that.isVertical() ? breadth(d.label) : that.options.margin[that.options.anchor]
-                , _y = that.isVertical() ? that.options.margin.top : breadth(d.label)
+              var _x = that.isVertical() ? that.breadthScale(d.label) : that.options.margin[that.options.anchor]
+                , _y = that.isVertical() ? that.options.margin.top : that.breadthScale(d.label)
               ;
               return 'translate(' + _x + ',' + _y +')';
             });
@@ -170,6 +129,8 @@ function BarGraph () {
             .style("fill", function(d) {
               return d.color;
             });
+
+
         }
         //
         // Single dimension just sets the barse
@@ -189,28 +150,38 @@ function BarGraph () {
         }
       });
 
-
-
-
-
-
-
-        //                   
-        //                   
-    // getLabelOptions: function () {
-
-    // },
-
-
-
-
-
-
+      this.onBarLayoutChange = function () {
+        this.draw();
+      };
 
       // sets a method to allow redrawing
       this.onAnchorChange = function () {
         this.draw();
       };
+    },
+
+    /**
+     * Sets the layout of the chart
+     * @param  {String} value - the type of layout to use allowed values = BARLAYOUT_GROUPED | BARLAYOUT_STACKED
+     * @return {Mixed}        - the value or chart
+     */
+    barLayout: function (value) {
+      if (!arguments.length) return this.options.barLayout;
+
+      value = String(value).toLowerCase();
+      // wrong value supplied
+      if (value !== this.BARLAYOUT_GROUPED &&
+          value !== this.BARLAYOUT_STACKED) {
+        console.error(value, 'is invalid. Only ', this.BARLAYOUT_STACKED, ' OR ', this.BARLAYOUT_GROUPED, ' allowed');
+        return this;
+      }
+      this.options.barLayout = value;
+      // redraw if the graph has been loaded
+      if (typeof this.onBarLayoutChange === 'function') {
+        this.onBarLayoutChange();
+      }
+
+      return this;
     },
 
     /**
@@ -268,8 +239,14 @@ function BarGraph () {
     }
   };
 
+
   chart = _.extend(chart, barOptions);
+  chart = _.extend(chart, barPositions);
+  chart = _.extend(chart, barDimensions);
   chart = _.extend(chart, labelOptions);
+// keep the options clean
+  chart.options = _.extend(chart.options, base.options);
+  base = _.omit(base, 'options');
   chart = _.extend(chart, base);
   return (chart.init());